Commands Registered by Proper Objects
