+++ fission/src/systems/match_mode/MatchMode.ts
@@ -10,6 +10,26 @@ import SimulationSystem from "../simulation/SimulationSystem"
 import { SoundPlayer } from "../sound/SoundPlayer"
 import { MatchModeType } from "./MatchModeTypes"
 import RobotDimensionTracker from "./RobotDimensionTracker"
+import CommandRegistry from "@/ui/components/CommandRegistry"
+import { globalAddToast, globalOpenPanel } from "@/ui/components/GlobalUIControls"
+
+// Register command: Toggle Match Mode
+CommandRegistry.get().registerCommand({
+	id: "toggle-match-mode",
+	label: "Toggle Match Mode",
+	description: "Toggle match mode, allowing you to simulate and run a full match.",
+	keywords: ["match", "mode", "start", "play", "game", "simulate", "toggle"],
+	perform: () => {
+		if (MatchMode.getInstance().isMatchEnabled()) {
+			MatchMode.getInstance().sandboxModeStart()
+			globalAddToast("info", "Match Mode Cancelled")
+		} else {
+			import("@/ui/panels/configuring/MatchModeConfigPanel").then(m => {
+				globalOpenPanel(m.default, undefined)
+			})
+		}
+	},
+})
 
 class MatchMode {
     private static _instance: MatchMode
