+++ fission/src/ui/panels/configuring/assembly-config/ConfigurePanel.tsx
@@ -31,6 +31,101 @@ import SequentialBehaviorsInterface from "./interfaces/SequentialBehaviorsInterf
 import SimulationInterface from "./interfaces/SimulationInterface"
 import ConfigureProtectedZonesInterface from "./interfaces/scoring/ConfigureProtectedZonesInterface"
 import ConfigureScoringZonesInterface from "./interfaces/scoring/ConfigureScoringZonesInterface"
+import CommandRegistry, { type CommandDefinition, type CommandProvider } from "@/ui/components/CommandRegistry"
+import { globalAddToast, globalOpenPanel } from "@/ui/components/GlobalUIControls"
+
+// Register command: Configure Assets (module-scope side effect)
+CommandRegistry.get().registerCommand({
+	id: "configure-assets",
+	label: "Configure Assets",
+	description: "Open the asset configuration panel.",
+	keywords: ["configure", "asset", "config"],
+	perform: () => import("./ConfigurePanel").then(m => globalOpenPanel(m.default, {})),
+})
+
+// Register command: Configure Robots (module-scope)
+CommandRegistry.get().registerCommand({
+	id: "configure-robots",
+	label: "Configure Robots",
+	description: "Open the configuration panel scoped to spawned robots.",
+	keywords: ["configure", "robot", "robots", "config"],
+	perform: () => {
+		const robots = World.sceneRenderer.mirabufSceneObjects.getRobots()
+		if (!robots || robots.length === 0) {
+			globalAddToast("warning", "No Robots", "No robots are currently spawned.")
+			return
+		}
+		import("./ConfigurePanel").then(m => globalOpenPanel(m.default, {
+			configurationType: "ROBOTS",
+			selectedAssembly: robots.length === 1 ? robots[0] : undefined,
+		}))
+	},
+})
+
+// Register dynamic provider: per-assembly configure/remove commands (module-scope)
+const provider: CommandProvider = () => {
+	if (!World.isAlive || !World.sceneRenderer) return []
+	const list: CommandDefinition[] = []
+
+	const robots = World.sceneRenderer.mirabufSceneObjects.getRobots() || []
+	for (const r of robots) {
+		const name = r.assemblyName || "Robot"
+		const nameTokens = String(name)
+			.split(/\s+|[-_]/g)
+			.filter(Boolean)
+		list.push({
+			id: `configure-robot-${r.id}`,
+			label: `Configure ${r.nameTag?.text()} (${name})`,
+			description: `Open configuration for robot ${r.nameTag?.text()} (${name}).`,
+			keywords: ["configure", "robot", ...nameTokens.map(t => t.toLowerCase())],
+			perform: () =>
+				import("./ConfigurePanel").then(m => globalOpenPanel(m.default, {
+					configurationType: "ROBOTS",
+					selectedAssembly: r,
+				})),
+		})
+		list.push({
+			id: `remove-robot-${r.id}`,
+			label: `Remove ${r.nameTag?.text()} (${name})`,
+			description: `Remove the robot ${r.nameTag?.text()} (${name}).`,
+			keywords: ["remove", "delete", "robot", ...nameTokens.map(t => t.toLowerCase())],
+			perform: () => {
+				World.sceneRenderer.removeSceneObject(r.id)
+			},
+		})
+	}
+
+	const field = World.sceneRenderer.mirabufSceneObjects.getField()
+	if (field) {
+		const name = field.assemblyName || "Field"
+		const nameTokens = String(name)
+			.split(/\s+|[-_]/g)
+			.filter(Boolean)
+		list.push({
+			id: `configure-field-${field.id}`,
+			label: `Configure ${name}`,
+			description: `Open configuration for field ${name}.`,
+			keywords: ["configure", "field", ...nameTokens.map(t => t.toLowerCase())],
+			perform: () =>
+				import("./ConfigurePanel").then(m => globalOpenPanel(m.default, {
+					configurationType: "FIELDS",
+					selectedAssembly: field,
+				})),
+		})
+		list.push({
+			id: `remove-field-${field.id}`,
+			label: `Remove ${name}`,
+			description: `Remove the field ${name}.`,
+			keywords: ["remove", "delete", "field", ...nameTokens.map(t => t.toLowerCase())],
+			perform: () => {
+				World.sceneRenderer.removeSceneObject(field.id)
+			},
+		})
+	}
+
+	return list
+}
+CommandRegistry.get().registerProvider(provider)
 
 interface ConfigInterfaceProps<T, P> {
     panel: UIScreen<T, P>
